requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
PyWavelets>=1.4.0
ruptures>=1.1.7
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pytest>=7.0.0
pyyaml>=6.0
tqdm>=4.65.0
pyarrow>=12.0.0
fastparquet>=2023.4.0
